requirements.txt
langchain>=0.2.0
langchain-ollama>=0.1.0
langgraph>=0.2.0
langchain-core>=0.2.0
transformers>=4.40.0
torch>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
